Edited manuscript code
Completely commented code for revised draft of manuscript submitted to AEM. Code is modular and broken into modules by identifying the figure/table for which the data is processed. Variables are defined multiple times, but commented out all but the first time. If the code is to be used in modules, the variables may need to be uncommented to function.

Original code written by Emily Ann McClure, https://github.com/mcclur51e